fix: move Eigen out of SpdlogFormatters
